bangalorecoders-main/code/ESG Survey Automation/survey_response/requirements.txt
jsonschema==4.17.3
langchain==0.0.154
PyPDF2==3.0.1
pymilvus==2.2.8
scikit-learn
python-dotenv==1.0.0
openai==0.27.7
Django==4.2.1
djangorestframework==3.14.0
tiktoken==0.4.0
reportlab
django-cors-headers